# openapi/bundled.yaml
openapi: 3.1.0
info:
  title: Chat API
  version: 0.1.0
  description: Slack-like communication application API
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/workspaces:
    get:
      operationId: listWorkspaces
      summary: List user workspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                required:
                  - workspaces
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createWorkspace
      summary: Create a new workspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Workspace id already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/workspaces/{id}:
    get:
      operationId: getWorkspace
      summary: Get workspace details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
                required:
                  - workspace
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateWorkspace
      summary: Update workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
                required:
                  - workspace
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteWorkspace
      summary: Delete workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Workspace deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/workspaces/public:
    get:
      operationId: listPublicWorkspaces
      summary: List all public workspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Public workspaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPublicWorkspacesResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/workspaces/{id}/join:
    post:
      operationId: joinPublicWorkspace
      summary: Join a public workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/workspaces/{id}/members:
    get:
      operationId: listMembers
      summary: List workspace members
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMembersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addMemberByEmail
      summary: Add member to workspace by email
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/workspaces/{id}/members/{userId}:
    patch:
      operationId: updateMemberRole
      summary: Update member role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Member role updated
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeMember
      summary: Remove member from workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - member
          default: member
      required:
        - email
...
    CreateWorkspaceRequest:
      type: object
      properties:
        id:
          type: string
          minLength: 3
          maxLength: 12
          pattern: ^[a-z0-9][a-z0-9-]*[a-z0-9]$
        name:
          type: string
          minLength: 1
        description:
          type: string
        iconUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
      required:
        - id
        - name
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ListMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
      required:
        - members
...
    ListPublicWorkspacesResponse:
      type: object
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/PublicWorkspaceItem'
      required:
        - workspaces
...
    MemberInfo:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - owner
            - admin
            - member
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - email
        - displayName
        - role
        - joinedAt
...
    PublicWorkspaceItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        iconUrl:
          type: string
          nullable: true
        memberCount:
          type: integer
          minimum: 0
        isJoined:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - memberCount
        - isJoined
        - createdAt
...
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - admin
            - member
      required:
        - role
...
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        iconUrl:
          type: string
          format: uri
          maxLength: 2048
        isPublic:
          type: boolean
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          description: slug identifier (3-12 chars, lowercase, digits, hyphen)
        name:
          type: string
        description:
          type: string
          nullable: true
        iconUrl:
          type: string
          nullable: true
        isPublic:
          type: boolean
        role:
          type: string
          enum:
            - owner
            - admin
            - member
            - guest
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdBy
        - createdAt
...